Files modified: 2
Modified 2 files:geom/color/types/HSV.swiftgeom/color/utils/ColorParser.swift

+++ geom/color/types/HSV.swift
@@ -1,9 +1,9 @@
 import Foundation
 class HSV {
-	var h:UInt/*0-240*/
-	var s:UInt/*0-240*/
-	var v:UInt/*0-240*/
-	init(_ h:UInt = 0, _ s:UInt = 0, _ v:UInt = 0) {
+	var h:CGFloat/*0-240*/
+	var s:CGFloat/*0-240*/
+	var v:CGFloat/*0-240*/
+	init(_ h:CGFloat = 0, _ s:CGFloat = 0, _ v:CGFloat = 0) {
 		self.h = h
 		self.s = s
 		self.v = v
